+++ Optuna/Study/Study.cs
@@ -136,14 +136,15 @@ namespace Optuna.Study
             return studySummaries;
         }
 
-        public static dynamic CreateStudy(dynamic optuna, string studyName, dynamic sampler, string[] directions, dynamic storage, bool loadIfExists = true)
+        public static dynamic CreateStudy(dynamic optuna, string studyName, dynamic sampler, string[] directions, dynamic storage, dynamic pruner, bool loadIfExists = true)
         {
             dynamic study = optuna.create_study(
                 sampler: sampler,
                 directions: directions,
                 storage: storage,
                 study_name: studyName,
-                load_if_exists: loadIfExists
+                load_if_exists: loadIfExists,
+                pruner: pruner
             );
             // for escape exception in Brute Force sampler
             // Study.stop() method throws exception when in_optimize_loop is false