+++ Tunny.Core/Settings/Optimize.cs
@@ -1,5 +1,4 @@
 using Tunny.Core.TEnum;
-using Tunny.Settings.Sampler;
 
 namespace Tunny.Core.Settings
 {