+++ Tunny.Core/Util/TEnvVariables.cs
@@ -10,6 +10,7 @@ namespace Tunny.Core.Util
         public static string DefaultStoragePath { get; } = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.Desktop), "fish.log");
         public static string TunnyEnvPath { get; } = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.UserProfile), ".tunny_env");
         public static string LogPath { get; } = Path.Combine(TunnyEnvPath, "logs");
+        public static string QuitFishingPath { get; } = Path.Combine(TunnyEnvPath, "quit.fishing");
         public static string DesignExplorerPath { get; } = Path.Combine(TunnyEnvPath, "TT-DesignExplorer");
         public static string OptimizeSettingsPath { get; } = Path.Combine(TunnyEnvPath, "settings.json");
         public static string PythonPath { get; } = Path.Combine(TunnyEnvPath, "python");